feat+fix: implemented calculator buttons logic and state handling + fixed floating point precision issue

+++ app/src/main/java/com/mx3/thirdwayvsimplecalculator/data/model/Operator.java
@@ -1,5 +1,7 @@
 package com.mx3.thirdwayvsimplecalculator.data.model;
 
+import java.math.BigDecimal;
+
 public abstract class Operator {
 
     // Constructor
@@ -7,11 +9,12 @@ public abstract class Operator {
     public Operator() {
     }
 
+
     // Other abstract methods
 
-    public abstract Float execute(Operand first, Operand second);
+    public abstract BigDecimal execute(Operand first, Operand second);
 
-    public abstract Float reverse(Operand first, Operand second);
+    public abstract BigDecimal reverse(Operand first, Operand second);
 
     public abstract char getChar();
 }
